pyproject.toml
[tool.poetry]
name = "codeshift.az"
version = "0.1.0"
description = ""
authors = ["Tarlan Soltanov <tarlansoltanov@outlook.com>"]
readme = "README.md"
packages = [{ include = "server" }]

[tool.poetry.dependencies]
python = "^3.10"

python-decouple = "^3.8"
colorlog = "^6.8.0"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
pillow = "^10.2.0"

django = "^5.0"
django-split-settings = "^1.2.0"
django-health-check = "^3.17.0"
django-cors-headers = "^4.3.1"
django-cleanup = "^8.0.0"

django-rosetta = "^0.9.9"
django-modeltranslation = "^0.18.11"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

faker = "^22.0.0"
freezegun = "^1.4.0"

pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"

django-debug-toolbar = "^4.2.0"

[tool.black]
line-length = 119
target-version = ['py310']

[tool.isort]
py_version = 310
profile = "black"
line_length = 119
order_by_type = false

[tool.flake8]
max-line-length = 119
ignore = ["A003", "W503"]

[tool.pytest.ini_options]
env = ["DJANGO_ENV = test"]
DJANGO_SETTINGS_MODULE = "server.settings"
filterwarnings = ["ignore::DeprecationWarning"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
